==> 1_dataloading/labelled_reddit/1_load_bigquery_labelled_reddit.ipynb <==
import csv
import pandas as pd

from google.cloud import bigquery

# set GCP credentials
import os
os.environ["GOOGLE_APPLICATION_CREDENTIALS"] = "/Users/Paul/Downloads/gcp-key.json"
client = bigquery.Client()

## Download Pushshift data via BigQuery

%%time 

for year in [2019]:
    for month in ['07', '08', '09', '10', '11', '12']: #'01', '02', '03', '04', '05', '06', 

        print(f"{year}_{month}")

        query_job = client.query(
            f"""
            SELECT created_utc, body, subreddit
            FROM `fh-bigquery.reddit_comments.{year}_{month}`
            WHERE subreddit = 'ChapoTrapHouse' OR subreddit = 'Conservative' OR subreddit = 'Libertarian' OR subreddit = 'The_Donald' OR subreddit = 'politics'
            """
        )

        results = query_job.result()  # Waits for job to complete.
        print('\t queried results')

        # write result to df
        results_df = pd.DataFrame({'created_utc': row.created_utc, 'body': row.body[:1024], 'subreddit': row.subreddit} for row in results)
        print('\t wrote to df')

        # export df to csv
        results_df.to_csv(f'../../0_data/raw/labelled_reddit/politics_{year}_{month}.csv', index=False)
        print('\t exported to csv')
        print()
