==> 1_dataloading/labelled_reddit/1_load_pushshift_labelled_reddit.ipynb <==
import datetime as dt
import csv
import pandas as pd
import os

from psaw import PushshiftAPI

# initialise API
api = PushshiftAPI()

## Load Comments via Pushshift API
Pushshift BigQuery does not provide data for after December 2019, so we query comments for Jan and Feb 2020 via the regular API

%%time

# initialise empty df for writing into
sample_df = pd.DataFrame(columns = ['created_utc', 'body', 'subreddit'])


for subreddit in ['politics']:

    print(subreddit)
    
    for year, month in [(2020, 1)]:
        print(year, month)

        start_time_epoch = int(dt.datetime(year = year, month = month, day = 1, hour = 0, minute = 0 ).timestamp())
        end_time_epoch = int(dt.datetime(year = year, month = month+1, day = 1, hour = 0, minute = 0 ).timestamp())

        gen = api.search_comments(before = end_time_epoch, after = start_time_epoch,
                                  subreddit = subreddit,
                                  filter = ['body', 'subreddit'])
        print(' queried results')

        # write result to df
        results_df = pd.DataFrame({'created_utc': row.created_utc, 'body': row.body, 'subreddit': row.subreddit} for row in gen)
        print(' wrote to df')

        # export df to csv
        results_df.to_csv(f'../../0_data/raw/labelled_reddit/sub_{subreddit}_{year}_0{month}.csv', index=False)
        print(' exported to csv')
        print()                              

## Combine subreddit dfs into single df for Jan and Feb 2020

directory = '../../0_data/raw/labelled_reddit'

for year, month in [(2020, 1), (2020, 2)]:
    
    print(year, month)
    
    export_df = pd.DataFrame(columns = ["created_utc", "body", "subreddit"])
    
    for filename in os.listdir(directory):
        if "sub_" in filename and filename.endswith(f"{year}_0{month}.csv"):
            print(' ', filename)
            export_df = export_df.append(pd.read_csv(os.path.join(directory, filename)))
            print(' ', export_df.shape)
    export_df.sample(frac=1, random_state=123).to_csv(os.path.join(directory, f"politics_{year}_0{month}.csv"), index = False)
    
    print()

## Number of Posts per Month by Subreddit

%%time

# number of posts per month by subreddit

post_counts = {}

for subreddit in ['politics', 'chapotraphouse', 'the_donald', 'libertarian', 'conservative']:
    
    post_counts[subreddit] = {}
    
    print(subreddit.upper())
    
    for year in [2017, 2018, 2019]:
        
        for month in range(1,13):
            
            start_time_epoch = int(dt.datetime(year = year, month = month, day = 1).timestamp())
            if month < 12:
                end_time_epoch = int(dt.datetime(year = year, month = month + 1, day = 1).timestamp())
            else:
                end_time_epoch = int(dt.datetime(year = year, month = month, day = 31, hour=23, minute=59).timestamp())

            gen = api.search_comments(before = end_time_epoch, after = start_time_epoch, subreddit = subreddit)
            thing = next(gen)
            post_counts[subreddit][f'{year}-{month}'] = api.metadata_['total_results']
            print(f'{year}-{month}\t', api.metadata_['total_results'])
                                         
    print()

%%time

# number of posts per month by subreddit

post_counts = {}

for subreddit in ['politics', 'chapotraphouse', 'the_donald', 'libertarian', 'conservative']:
    
    post_counts[subreddit] = {}
    
    print(subreddit.upper())
    
    for year in [2020]:
        
        for month in range(1,3):
            
            start_time_epoch = int(dt.datetime(year = year, month = month, day = 1).timestamp())
            if month < 12:
                end_time_epoch = int(dt.datetime(year = year, month = month + 1, day = 1).timestamp())
            else:
                end_time_epoch = int(dt.datetime(year = year, month = month, day = 31, hour=23, minute=59).timestamp())

            gen = api.search_comments(before = end_time_epoch, after = start_time_epoch, subreddit = subreddit)
            thing = next(gen)
            post_counts[subreddit][f'{year}-{month}'] = api.metadata_['total_results']
            print(f'{year}-{month}\t', api.metadata_['total_results'])
                                         
    print()

pd.DataFrame.from_dict(post_counts)
