==> 1_dataloading/labelled_reddit/3_total_splits_labelled_reddit.ipynb <==
# Create Total Splits from Monthly Splits - Labelled Reddit Data

import csv
import random
import os
import pandas as pd
from sklearn.model_selection import train_test_split

# Helper functions

def load_from_csv(filepath, month_sample_size):
    import_df = pd.read_csv(filepath)
    sample_df, _ = train_test_split(import_df, train_size = month_sample_size, stratify=import_df.label, random_state=123)
    return sample_df

def create_sample(source_dir, mode, size, label):
    
    comment_df = pd.DataFrame(columns = ['clean_text', 'label'])
    
    for filename in sorted(os.listdir(source_dir)):
        if mode in filename and filename.endswith("40k.csv" if mode=='train' else "10k.csv"):
            print(f"  loading subsample from {filename}")
            comment_df = comment_df.append(load_from_csv(os.path.join(source_dir, filename), month_sample_size = int(round(size/36, 0)+1))) # divide by 36 to stratify across months
    
    export_df, _ = train_test_split(comment_df, train_size = size, stratify = comment_df.label, random_state = 123)
    
    export_df.to_csv(f'../../0_data/clean/labelled_reddit/total/{mode}_rand_{label}.csv', index=False)

# Creating random training and test sets

%%time

# create training sets
mode = 'train'

for size, label in [(2000, '2k'), (5000, '5k'), (10000, '10k')]:
    print(f"creating random {mode} set, size {label}")
    create_sample(source_dir = '../../0_data/clean/labelled_reddit/month_splits', mode = mode, size = size, label = label)

%%time

# create test sets
mode = 'test'

for size, label in [(5000, '5k'), (10000, '10k'), (20000, '20k'), (40000, '40k')]:
    print(f"creating random {mode} set, size {label}")
    create_sample(source_dir = '../../0_data/clean/labelled_reddit/month_splits', mode = mode, size = size, label = label)
