==> 1_dataloading/labelled_reddit/4_labelled_to_unlabelled_test.ipynb <==
import csv
import os
import pandas as pd

source_dir = "../../0_data/clean/labelled_reddit/month_splits"


for filename in sorted(os.listdir(source_dir)):
        if filename.startswith("test_") and filename.endswith("_5k.csv"):
            print(filename)
            import_df = pd.read_csv(os.path.join(source_dir,filename))
            
            with open(f'../../0_data/clean/unlabelled_reddit/politics_test/{filename.strip(".csv")}.txt', 'w') as write_obj:
                for text in import_df.clean_text.sample(frac=1, random_state=123):
                    write_obj.write(text + "\n \n")
